moved validation methods to validate-event file
