+++ src/data-types/interfaces.ts
@@ -64,7 +64,7 @@ export interface TxMAEgressExtensions {
 export interface TxMAIngressEvent {
   event_name: string;
   timestamp: number;
-  event_timestamp_ms: number;
+  event_timestamp_ms?: number;
   component_id: string;
   user: TxmaUser;
   extensions?: IngressEventExtension;
