+++ src/services/validate-event.ts
@@ -1,10 +1,12 @@
-import { TxMAIngressEvent } from '../data-types/interfaces';
+import { DynamoDBStateResult, TxMAIngressEvent } from '../data-types/interfaces';
 import logger from '../commons/logger';
 import { logAndPublishMetric } from '../commons/metrics';
-import { LOGS_PREFIX_SENSITIVE_INFO, MetricNames } from '../data-types/constants';
+import { EventsEnum, LOGS_PREFIX_SENSITIVE_INFO, MetricNames } from '../data-types/constants';
 import { ValidationError } from '../data-types/errors';
 import { compileSchema } from '../commons/compile-schema';
 import { TxMAIngress } from '../data-types/schemas';
+import { getCurrentTimestamp } from '../commons/get-current-timestamp';
+import { sendAuditEvent } from './send-audit-events';
 
 const validateInterventionDataInput = compileSchema(TxMAIngress);
 
@@ -13,7 +15,7 @@ const validateInterventionDataInput = compileSchema(TxMAIngress);
  *
  * @param interventionRequest - the TxMA request
  */
-export function validateEvent(interventionRequest: TxMAIngressEvent): void {
+export function validateEventAgainstSchema(interventionRequest: TxMAIngressEvent): void {
   if (!validateInterventionDataInput({ event: interventionRequest })) {
     logger.debug(`${LOGS_PREFIX_SENSITIVE_INFO} Event has failed schema validation.`, {
       validationErrors: validateInterventionDataInput.errors,
@@ -34,4 +36,50 @@ export function validateInterventionEvent(interventionRequest: TxMAIngressEvent)
     logAndPublishMetric(MetricNames.INVALID_EVENT_RECEIVED);
     throw new ValidationError('Invalid intervention event.');
   }
+}
+
+export function validateLevelOfConfidence(intervention: EventsEnum, event: TxMAIngressEvent) {
+  if (intervention === EventsEnum.IPV_IDENTITY_ISSUED && event.extensions?.levelOfConfidence !== 'P2') {
+    logger.warn(`Received interventions has low level of confidence: ${event.extensions?.levelOfConfidence}`);
+    logAndPublishMetric(MetricNames.CONFIDENCE_LEVEL_TOO_LOW);
+    throw new ValidationError('Received intervention has low level of confidence.');
+  }
+}
+
+export async function validateEventIsNotInFuture(intervention: EventsEnum, event: TxMAIngressEvent) {
+  const eventTimestampInMs = event.event_timestamp_ms ?? event.timestamp * 1000;
+  const now = getCurrentTimestamp().milliseconds;
+  if (now < eventTimestampInMs) {
+    logger.debug(`Timestamp is in the future (sec): ${eventTimestampInMs}.`);
+    logAndPublishMetric(MetricNames.INTERVENTION_IGNORED_IN_FUTURE);
+    await sendAuditEvent('AIS_INTERVENTION_IGNORED_IN_FUTURE', event.user.user_id, {
+      intervention,
+      reason: 'received event is in the future',
+      appliedAt: undefined,
+    });
+    throw new Error('Event is in the future. It will be retried');
+  }
+}
+
+export async function validateEventIsNotStale(
+  intervention: EventsEnum,
+  event: TxMAIngressEvent,
+  itemFromDB?: DynamoDBStateResult,
+) {
+  const eventTimestampInMs = event.event_timestamp_ms ?? event.timestamp * 1000;
+  if (!isEventAfterLastEvent(eventTimestampInMs, itemFromDB?.sentAt, itemFromDB?.appliedAt)) {
+    logger.warn('Event received predates last applied event for this user.');
+    logAndPublishMetric(MetricNames.INTERVENTION_EVENT_STALE);
+    await sendAuditEvent('AIS_INTERVENTION_IGNORED_STALE', event.user.user_id, {
+      intervention,
+      appliedAt: undefined,
+      reason: 'Received intervention predates latest applied intervention',
+    });
+    throw new ValidationError('Event received predates last applied event for this user.');
+  }
+}
+
+function isEventAfterLastEvent(eventTimeStamp: number, sentAt?: number, appliedAt?: number) {
+  const latestIntervention = sentAt ?? appliedAt ?? 0;
+  return eventTimeStamp > latestIntervention;
 }
